+++ src/index.html
@@ -55,5 +55,6 @@
     <include src="./partials/marckups/button-markups.html"></include>
 
     <script type="module" src="./index.js"></script>
+    <script type="module" src="./js/fetch/searchByName.js"></script>
   </body>
 </html>
