+++ src/index.js
@@ -12,6 +12,6 @@ import './js/team-modal/team-modal';
 import './js/trailer/trailer';
 
 import './js/theme-switcher/theme-switcher';
-
+import './js/fetch/pagination/btn_up';
 import './js/fetch/pagination/pagination';
-import './js/fetch/searchByName';
+// import './js/fetch/searchByName'; відразу підвязала в index.html
