+++ src/js/fetch/fetch.js
@@ -11,7 +11,7 @@ class MoviesApiService {
     try {
       const url = `${BASE_URL}trending/movie/week?api_key=${API_KEY}&language=en-US&page=${this.page}`;
       const response = await axios.get(url);
-      console.log(response);
+      // console.log(response);
       return response.data;
     } catch (error) {
       console.error(error);
